+++ src/main/java/org/observertc/webrtc/observer/connectors/encoders/EncoderConfig.java
@@ -1,5 +1,6 @@
 package org.observertc.webrtc.observer.connectors.encoders;
 
+import org.observertc.webrtc.observer.connectors.MessageFormat;
 import org.observertc.webrtc.observer.connectors.encoders.avro.AvroEncoder;
 
 import java.util.Map;
@@ -7,7 +8,7 @@ import java.util.Map;
 public class EncoderConfig {
 
     public String type = AvroEncoder.class.getSimpleName();
-
+    public MessageFormat format = MessageFormat.BYTES;
     public Map<String, Object> config;
 
 }