+++ CHANGELOG.md
@@ -4,8 +4,8 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 ## 0.7.3 (H)
 
 ### Added
- * Add Jackson Encoder to the encoder type, in order to transform avro reports to json format
- * Add `messageFormat` field to the LoggerSink configuration to support printing and accepting json type encoded messages.
+ * Add Json Encoder to the encoder type, in order to transform avro reports to json format
+ * Add `format` to encoder config to indicate the output format (`bytes`, `object`) depending on the sink requirements
  * Added exposed metric `call_durations_in_mins` reports the duration of an ended call at the time it is ended
  * Add marker field to the PeerConnectionSample schema
  * Added new address for pc samples `/pcsamples/{serviceUUID}/{mediaUnitId}`